+++ test/jest/controls/control.date.test.js
@@ -246,6 +246,7 @@ describe("control.date", () => {
     const onSubmit = jest.fn();
     form.$onSubmit = onSubmit;
     el.$options.name = "testDate";
+    h.setupLayout(el, { h: 20, w: 100 });
     await h.wait(1);
 
     HTMLInputElement.prototype.focus.call(el.$refInput);